requirements.txt
gradio>=4.27.0
python-telegram-bot==20.8
